refactor: refine language around transaction signing
