+++ src/app/pages/transaction-request/components/contract-call-details/contract-call-details.spec.tsx
