+++ src/app/pages/transaction-request/components/contract-call-details/contract-call-details.tsx
@@ -4,8 +4,8 @@ import { Stack, color } from '@stacks/ui';
 import { useExplorerLink } from '@app/common/hooks/use-explorer-link';
 import { Divider } from '@app/components/divider';
 import { Title } from '@app/components/typography';
-import { ContractPreview } from '@app/pages/sign-transaction/components/contract-preview';
-import { AttachmentRow } from '@app/pages/sign-transaction/components/attachment-row';
+import { ContractPreview } from '@app/pages/transaction-request/components/contract-preview';
+import { AttachmentRow } from '@app/pages/transaction-request/components/attachment-row';
 import { useTransactionRequestState } from '@app/store/transactions/requests.hooks';
 
 import { FunctionArgumentsList } from './function-arguments-list';
