+++ src/app/pages/transaction-request/components/contract-call-details/function-argument-item.tsx
@@ -1,6 +1,6 @@
 import { deserializeCV, cvToString, getCVTypeString } from '@stacks/transactions';
 
-import { Row } from '@app/pages/sign-transaction/components/row';
+import { Row } from '@app/pages/transaction-request/components/row';
 import { useContractFunction } from '@app/query/contract/contract.hooks';
 
 interface FunctionArgumentProps {
