+++ src/app/pages/transaction-request/components/contract-call-details/function-arguments-list.tsx
