+++ src/app/pages/transaction-request/components/contract-deploy-details/contract-deploy-details.tsx
@@ -6,9 +6,9 @@ import { Prism } from '@app/common/clarity-prism';
 import { useWallet } from '@app/common/hooks/use-wallet';
 import { Caption, Title } from '@app/components/typography';
 import { Divider } from '@app/components/divider';
-import { ContractPreview } from '@app/pages/sign-transaction/components/contract-preview';
-import { Row } from '@app/pages/sign-transaction/components/row';
-import { AttachmentRow } from '@app/pages/sign-transaction/components/attachment-row';
+import { ContractPreview } from '@app/pages/transaction-request/components/contract-preview';
+import { Row } from '@app/pages/transaction-request/components/row';
+import { AttachmentRow } from '@app/pages/transaction-request/components/attachment-row';
 
 function ContractCodeSection(): JSX.Element | null {
   const pendingTransaction = usePendingTransaction();
