+++ src/app/pages/transaction-request/components/contract-preview.tsx
