+++ src/app/pages/transaction-request/components/post-condition-mode-warning.tsx
