+++ src/app/pages/transaction-request/components/post-conditions/post-condition-item.tsx
