+++ src/app/pages/transaction-request/components/post-conditions/post-conditions.tsx
