+++ src/app/pages/transaction-request/components/post-conditions/stx-post-condition.tsx
