+++ src/app/pages/transaction-request/components/stx-transfer-details/stx-transfer-details.spec.tsx
