+++ src/app/pages/transaction-request/components/transaction-error/error-message.tsx
