+++ src/app/pages/transaction-request/components/transaction-error/transaction-error.spec.tsx
