+++ src/content-scripts/content-script.ts
@@ -87,7 +87,7 @@ document.addEventListener(DomEventName.authenticationRequest, ((
 // Listen for a CustomEvent (transaction request) coming from the web app
 document.addEventListener(DomEventName.transactionRequest, ((event: TransactionRequestEvent) => {
   forwardDomEventToBackground({
-    path: RouteUrls.Transaction,
+    path: RouteUrls.TransactionRequest,
     payload: event.detail.transactionRequest,
     urlParam: 'request',
     method: ExternalMethods.transactionRequest,
